fix(iam): resolve team ID via /identity/whoami instead of SAML config
cx iam users search (and get/create/update/set-status) was failing for
any API key lacking the SAML scope, and silently returning an empty list
for all other keys because the search endpoint requires pageSize to
return results.

Two bugs fixed:
1. Team ID resolution: was calling GET /aaa/team-saml/v1/configuration,
   which requires a SAML-specific scope unrelated to user management and
   fails outright on teams without SAML configured. Replaced with
   GET /identity/whoami, which every API key can read.
2. Missing pageSize: the /aaa/teams/v2/{id}/search endpoint silently
   returns an empty users array without a pageSize param. Now defaults
   to 300 when the caller does not specify one.

The whoami logic is extracted into a shared src/identity.rs module so
both the users and cases commands use a single implementation.
