+++ src/commands/users/mod.rs
@@ -7,7 +7,6 @@ use colored::Colorize;
 use serde_json::{json, Value};
 use toon_format::encode_default as toon_encode;
 
-use crate::commands::saml::api::SamlApi;
 use crate::config::OutputFormat;
 use crate::execution::{fan_out, ExecutionTarget};
 use crate::render;
@@ -46,24 +45,6 @@ fn read_from_file(path: &str) -> Result<Value> {
     Ok(serde_json::from_str(&raw)?)
 }
 
-async fn resolve_team_id(client: &crate::api_client::CxClient) -> anyhow::Result<String> {
-    let saml = SamlApi::new(client);
-    let config = saml.get_config().await.map_err(|e| {
-        let msg = e.to_string();
-        if msg.contains("Permission denied") || msg.contains("Authentication failed") {
-            anyhow::anyhow!(
-                "Cannot resolve team ID: API key lacks SAML scope (required by the users API)"
-            )
-        } else {
-            anyhow::anyhow!("Failed to resolve team ID from SAML config: {e}")
-        }
-    })?;
-    config
-        .team_id
-        .map(|id| id.to_string())
-        .ok_or_else(|| anyhow::anyhow!("SAML config returned no team ID"))
-}
-
 pub async fn run_search(
     targets: &[Arc<ExecutionTarget>],
     query: Option<&str>,
@@ -87,7 +68,7 @@ pub async fn run_search(
         let page_token = page_token.clone();
         async move {
             let api = UsersApi::new(&t.client);
-            let team_id = resolve_team_id(&t.client).await?;
+            let team_id = crate::identity::resolve_team_id(&t.client).await?;
             let team_id = team_id.as_str();
             let mut params: Vec<(&str, String)> = Vec::new();
             if let Some(ref q) = query {
@@ -96,19 +77,17 @@ pub async fn run_search(
             if let Some(ref s) = status {
                 params.push(("status", s.clone()));
             }
-            if let Some(ref ps) = page_size {
-                params.push(("pageSize", ps.clone()));
-            }
+            // pageSize is required — the server returns an empty list without it.
+            params.push((
+                "pageSize",
+                page_size.clone().unwrap_or_else(|| "300".to_string()),
+            ));
             if let Some(ref pt) = page_token {
                 params.push(("pageToken", pt.clone()));
             }
             let params_refs: Vec<(&str, &str)> =
                 params.iter().map(|(k, v)| (*k, v.as_str())).collect();
-            if params_refs.is_empty() {
-                Ok(api.search(team_id).await?)
-            } else {
-                Ok(api.search_with_params(team_id, &params_refs).await?)
-            }
+            Ok(api.search_with_params(team_id, &params_refs).await?)
         }
     })
     .await;
@@ -175,7 +154,7 @@ pub async fn run_get(
         let user_id = user_id.clone();
         async move {
             let api = UsersApi::new(&t.client);
-            let team_id = resolve_team_id(&t.client).await?;
+            let team_id = crate::identity::resolve_team_id(&t.client).await?;
             let team_id = team_id.as_str();
             Ok(api.get(team_id, &user_id).await?)
         }
@@ -226,7 +205,7 @@ pub async fn run_create(
         let body = body.clone();
         async move {
             let api = UsersApi::new(&t.client);
-            let team_id = resolve_team_id(&t.client).await?;
+            let team_id = crate::identity::resolve_team_id(&t.client).await?;
             let team_id = team_id.as_str();
             api.create(team_id, &body).await?;
             Ok(())
@@ -264,7 +243,7 @@ pub async fn run_update(
         let body = body.clone();
         async move {
             let api = UsersApi::new(&t.client);
-            let team_id = resolve_team_id(&t.client).await?;
+            let team_id = crate::identity::resolve_team_id(&t.client).await?;
             let team_id = team_id.as_str();
             Ok(api.update(team_id, &body).await?)
         }
@@ -327,7 +306,7 @@ pub async fn run_set_status(
         let body = body.clone();
         async move {
             let api = UsersApi::new(&t.client);
-            let team_id = resolve_team_id(&t.client).await?;
+            let team_id = crate::identity::resolve_team_id(&t.client).await?;
             let team_id = team_id.as_str();
             api.update_statuses(team_id, &body).await?;
             Ok(())
