+++ src/identity.rs
@@ -1,0 +1,52 @@
+use anyhow::anyhow;
+use serde::Deserialize;
+
+use crate::api_client::CxClient;
+
+const WHOAMI_BASE: &str = "/identity/whoami";
+
+/// Response from `GET /identity/whoami`. Identifies the team (and caller) that
+/// the current API key belongs to. Payload uses snake_case field names.
+#[derive(Debug, Deserialize)]
+pub struct Whoami {
+    pub team_id: Option<i64>,
+    pub team_name: Option<String>,
+    pub user_name: Option<String>,
+}
+
+/// Resolve the team ID for the current API key via `GET /identity/whoami`.
+///
+/// Preferred over the SAML config endpoint because `/identity/whoami` is
+/// readable by every API key regardless of scopes, and works on teams that
+/// have not configured SAML.
+pub async fn resolve_team_id(client: &CxClient) -> anyhow::Result<String> {
+    let whoami: Whoami = client
+        .get(WHOAMI_BASE, &[])
+        .await
+        .map_err(|e| anyhow!("failed to resolve team ID via /identity/whoami: {e:#}"))?;
+    whoami
+        .team_id
+        .map(|id| id.to_string())
+        .ok_or_else(|| anyhow!("/identity/whoami returned no team_id"))
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn deserialize_whoami() {
+        let json = serde_json::json!({ "team_id": 53623, "team_name": "acme", "user_name": "alice@example.com" });
+        let w: Whoami = serde_json::from_value(json).unwrap();
+        assert_eq!(w.team_id, Some(53623));
+        assert_eq!(w.team_name.as_deref(), Some("acme"));
+    }
+
+    #[test]
+    fn deserialize_whoami_minimal() {
+        let json = serde_json::json!({ "team_id": 12345 });
+        let w: Whoami = serde_json::from_value(json).unwrap();
+        assert_eq!(w.team_id, Some(12345));
+        assert!(w.team_name.is_none());
+    }
+}
