+++ tests/users/main.rs
@@ -2,7 +2,7 @@
 mod common;
 
 use serde_json::json;
-use wiremock::matchers::{method, path};
+use wiremock::matchers::{method, path, query_param};
 use wiremock::{Mock, MockServer, ResponseTemplate};
 
 use coralogix_cli::commands::users::run_search;
@@ -13,10 +13,8 @@ async fn search_users_from_mock() {
     let server = MockServer::start().await;
 
     Mock::given(method("GET"))
-        .and(path("/mgmt/openapi/5/aaa/team-saml/v1/configuration"))
-        .respond_with(
-            ResponseTemplate::new(200).set_body_json(json!({"teamId": 12345, "enabled": false})),
-        )
+        .and(path("/identity/whoami"))
+        .respond_with(ResponseTemplate::new(200).set_body_json(json!({ "team_id": 12345 })))
         .mount(&server)
         .await;
 
@@ -28,6 +26,7 @@ async fn search_users_from_mock() {
 
     Mock::given(method("GET"))
         .and(path("/mgmt/openapi/5/aaa/teams/v2/12345/search"))
+        .and(query_param("pageSize", "300"))
         .respond_with(ResponseTemplate::new(200).set_body_json(&body))
         .expect(1)
         .mount(&server)
